Tune the movebase param
The local planning goal was deviating from the global path

I tuned the path and goal distance bias to make the robot follows the
global path more. However, if the values are too high, then the cost of
staying will be lower than moving to goal. If the global path value is
too high, it will be difficult for the robot to go to a pose of local goal.
This will cause the robot to spin more on certain goal pose.

https://arxiv.org/pdf/1706.09068.pdf#:~:text=path%20distance%20bias%20is%20the,local%20goal%2C%20with%20whatever%20path.